moving service ui logic to its own class
