+++ Assets/BossRoom/Scenes/Startup.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7b2c33d4681d53ec62b32b7ff2c81123b4457d8dcd8d0b6d643097e74ae36856
-size 35006
+oid sha256:22939a0b19a2e47d1c43cf76d229f2a99bcc8511f03020367645a163b6ab2fba
+size 35402
