+++ Assets/BossRoom/Scripts/Client/UI/Lobby/LobbyUIMediator.cs
@@ -26,18 +26,15 @@ namespace Unity.Multiplayer.Samples.BossRoom.Visual
         NameGenerationData m_NameGenerationData;
         GameNetPortal m_GameNetPortal;
         ClientGameNetPortal m_ClientNetPortal;
-        private IDisposable m_Subscriptions;
 
         [Inject]
         void InjectDependenciesAndInitialize(
             LobbyServiceFacade lobbyServiceFacade,
-            IPublisher<UnityServiceErrorMessage> unityServiceErrorMessagePublisher,
             LocalLobbyUser localUser,
             LocalLobby localLobby,
             NameGenerationData nameGenerationData,
             GameNetPortal gameNetPortal,
-            ClientGameNetPortal clientGameNetPortal,
-            ISubscriber<UnityServiceErrorMessage> serviceError
+            ClientGameNetPortal clientGameNetPortal
         )
         {
             m_NameGenerationData = nameGenerationData;
@@ -50,24 +47,6 @@ namespace Unity.Multiplayer.Samples.BossRoom.Visual
             RegenerateName();
 
             m_ClientNetPortal.NetworkTimedOut += OnNetworkTimeout;
-            m_Subscriptions = serviceError.Subscribe(ServiceErrorHandler);
-        }
-
-        private void ServiceErrorHandler(UnityServiceErrorMessage error)
-        {
-            var errorMessage = error.Message;
-            if (error.AffectedService == UnityServiceErrorMessage.Service.Lobby)
-            {
-                if ((error.OriginalException as LobbyServiceException).Reason == LobbyExceptionReason.LobbyConflict)
-                {
-                    // LobbyConflict can have multiple causes. Let's add other solutions here if there's other situations that arise for this.
-                    errorMessage += "\nSee logs for possible causes and solution.";
-                    Debug.LogError($"Got service error {error.Message} with LobbyConflict. Possible conflict cause: Trying to play with two builds on the " +
-                              $"same machine. Please use command line arg '{ClientMainMenuState.AuthProfileCommandLineArg} someName' to set a different auth profile.\n");
-                }
-            }
-
-            PopupPanel.ShowPopupPanel("Service error", errorMessage);
         }
 
         void OnDestroy()
@@ -76,8 +55,6 @@ namespace Unity.Multiplayer.Samples.BossRoom.Visual
             {
                 m_ClientNetPortal.NetworkTimedOut -= OnNetworkTimeout;
             }
-
-            m_Subscriptions.Dispose();
         }
 
         //Lobby and Relay calls done from UI
