+++ Assets/BossRoom/Scripts/Client/UI/UIServices.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3baf99ef58264b4fb4bcec2f9ec2852
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
